Added softclipping function to filter out malformed alignments
GMAP before version 2020.10.12 has a bug that generates MD strings with asterisks for alignments that are both hardclipped (due to chromosome boundary) and softclipped on one end. These alignments generates error during variant calling. Version 2020.10.12 is not available at the moment through conda so we have to patch it for now.